Add rough port of dartsv library (#7)
This commit takes the dartsv library and repurposes it for use in Cashew
on the BCH network. This was done because it seems to be the only
Bitcoin-Cash compatible library that is MIT Licensed. The BitBox Lite
SDK has no license and is thus unsuitable for use.

The original source can be found here:
https://github.com/twostack/dartsv