Add a basic viewmodel for the existing UX and use provider (#27)
This commit introduces a ViewModel for the app. It allows components
and state to be tracked significantly easier than adding individual
variables and FutureBuilder across the app. Though, be careful to not
introduce globals into the ViewModel in the future. Information should
be as scoped as possible to allow for better unit testing.