Ensure Send Info screen isn't constantly redrawn (#35)
Because the flag to show the SendInfo screen was attached to the global
ViewModel, the SendInfo screen was redrawing whenever any of the other
viewmodel fields were being updated. (Which that screen does a lot of).
This was causing it to redraw and thus reset some internal state such as
where cursors were positioned in text fields.

This moves the state to a ValueNotifier within the specific Send screen
tab widget.