Fork AS1 libraries locally w/ serialization fix for integers
There is a bug in pointycastle ASN utils that causes BigInts to be
serialized incorrectly if the first byte has 0x80 set. (Indicating that
it is negative). Since this particular portion of the pointycastle lib
is relatively independet, and only used in a few places in our codebase,
I am temporarily forking it until we can get the fix upstreamed.