Redesign SendInfo screen to have a better input method for amounts
Implement a "calculator keyboard" and use it on the send information
data entry screen instead of a standard input box. Additionally, we
also get rid of the text entry field for an address and instead only
allow users to paste into it. There are several other misc things thrown
in with this redesign as well (e.g validation output that disables
the send button).