Refactor wallet initialization and move save/restore to viewmodel (#77)
This commit lifts the disk read/writes to the ViewModel (where it should
be), and changes the initialization flow so that the whole app is
reactive to changes on that model. Part of this is done by splitting out
the Send screen model into it's own entity, so that changes to that do
not cause a full app refresh. This way, only changes to the wallet/seed
cause this behavior.

I realize this change is rather huge, and probably should be swapped out
for several peices. However, I don't have time to split it up, and it's=
not immediately obvious how they could be split. There's essentially aborts
few things going on:

1. Lifitng up storage to the viewmodel.
2. Making the balance reactive so that a loading screen is not necessary
   and fact that the wallet isn't initialized does not explode the app
   while it's initializing (without a loading screen);
3. Splitting the WalletModel from the SendModel so that the WalletModel
   updates may be subscribed to. Changes to the seed will cause the
   entire app to refresh essentially.

Co-authored-by: Harry Barber <harrybarber@protonmail.com>