Fix transaction construction and wallet structure (#86)
The vault, and wallet, are not currently able to handle transaction
construction ideally. Neither can they handle a situation where a
transaction fails to broadcast and the UTXOs need to be-readded to the
vault. As such, this code refactors all of this from the ground up to
make everything work. However, some of the data structures are still
not ideal and should be reworked in the future.