Fix electrum ping response handling
Electrum returns a null result, and null error on `ping.` However,
dart completeError bans the return of nulls. Also, a null response
is not an error in this case. Thus, we just want to return null as
a good response except when `error` is explicitly set as a response
by the electrum server.