Generate keys on start rather than save them
It really does not take very long to generate child keys, assuming that
we have the xpriv key. This commit does not save/restore the keys and
instead generates them on load from the xpriv which is saved. This
allows us to avoid dealing with some state, which is generally always
a good thing.