Add HDPrivateKey tests and fix a bug with key derivation (#111)
Sometimes, sub-keys are encded into bigints of <32 bytes. In this case,
the automatic padding fails to be added correctly. Note: This is
currently wrong as it is using a broken version of encodeBigInt, which
we will fix in a later commit. However, attempting to fix this directly
resulted in other bugs we want to be sure we catch with the test.