Fix another issue with key encodings (#130)
In some cases, an xprivkey cannot be generated because the normal
serialization is less than 33 bytes. This means the CKDSerializer class
blows up, because it currently makes bad assumptions. The quick fix is
to move padding to 33 bytes into the CKD Serializer class. However, upon
further investigation, this class needs to be refactored itself.