Ensure we specify we need INTERNET permission in AndroidAmanifest
After many bugs fixed, the main issue with never finishing "Loading..."
turned out to be that we were not telling Android that we wanted to
access the internet. Alas, the other bugs were also real, and now the
software should be substantially more robust.

The dasterdly Loading... error gone (hopefully) once and for all.