Fix amount processing on the receive screen from the pay-to URI
Currently, there is a mismatch between the expectations on the send
and receive screen with relation to the amount in the URI. This commit
attempts to estimate if the amount when scanning a URI is in BCHA, or
in sats. It converts BCHA to sats for the purpose of sending.