Use oldest UTXOs first to avoid spending immature outputs
Currently, we select random UTXOs to spend. However, this causes
problems when spending from a wallet that you are mining to. This
patch changes the behavior to spend oldest coins first. Eventually,
we need to filter out immature outputs, but this is not trivial to do
with the current Fulcrum APIs and the data flows in Cashew.