Ensure app reconnects to electrum and refreshes balance when resumed
Add notification to the app to check when the app is resumed and re-
initialize the wallet when that occurs. This adds some additional code
to make sure that UTXOs are not added to the Utxo vault twice, as well
as ensuring that the electrum client properly reconnects when
disconnected.