Move lotus library tests into a proper subfolder of the test directory
In order to make tests run automatically from `flutter test` they need
to conform to the appropriate directory layout. Eventually, we want to
factor out the Lotus library into it's own published library, but for
now we will leave it as part of this project due to the need to iterate
on the APIs.

Additionally, this fixes privatekey serialization to WIFs (Wallet Import
Format), as well as the checksums on xpub serialization. Both of these
had been broken due to tests not running. Finally, a bitcoincash->lotus
find and replace in a past PR had caused addresses specified in tests to
have invalid checksums.